consoloidate all hybrid evil into one place
